Refactors auth form and login page
Improves code reusability by introducing a `TextLink` component and applying it to the auth form and login page.

This change replaces the inline styling of links with the `TextLink` component, promoting consistency and maintainability. It also adds `submitLoadingText` and `submitLabel` props to the `Form` component to allow for custom text on the submit button.